fix: rename plugin
